refactor: 자동차 이름을 포장하여 검증 역할 분리
