+++ README.md
@@ -44,4 +44,5 @@
 - [ ] 의미 있는 단위 테스트 고려하기
 - [ ] 역할이 많은 객체 분리하기
   - [x] Controller 에서 수행되는 도메인 생성 역할 분리
+  - [x] 문자열을 포장하여 관련 검증 역할 분리
 - [ ] 다양한 예외상황에 대해 대응하기
